Update Memisahkan logika player dan data room
