+++ src/services/PlayerDataServices.jsx
@@ -1,26 +1,32 @@
-// src/services/PlayerDataServices.js
+import {
+  database,
+  ref,
+  set,
+  get,
+  remove,
+  onValue,
+  runTransaction,
+} from "../firebaseConfig";
+import { resetRoom, getRoomPlayerCount } from "./roomDataServices";
 
-import { database, ref, set, get, remove, onValue, update } from "../firebaseConfig";
-
-export const checkRoomType = async (topicID, gameID, roomID) => {
-  const roomRef = ref(database, `rooms/${topicID}/${gameID}/${roomID}`);
-  const snapshot = await get(roomRef);
-  return {
-    exists: snapshot.exists(),
-    isSinglePlayer: snapshot.val()?.isSinglePlayer || false
-  };
-};
-
-export const getCurrentPlayers = async (topicID, gameID, roomID) => {
-  if (roomID === "room5") return [];
-  
-  const playersRef = ref(database, `rooms/${topicID}/${gameID}/${roomID}/players`);
-  const snapshot = await get(playersRef);
-  return snapshot.val() ? Object.values(snapshot.val()) : [];
-};
-
-export const roomsParticipation = async (topicID, gameID, roomID, user, isJoining, userPhoto) => {
-  if (roomID === "room5") return;
+/**
+ * Mengelola proses pemain bergabung atau keluar dari room.
+ * @param {string} topicID - ID topik.
+ * @param {string} gameID - ID game.
+ * @param {string} roomID - ID room.
+ * @param {object} user - Objek user yang berisi UID dan detail lainnya.
+ * @param {boolean} isJoining - Menentukan apakah pemain bergabung (true) atau keluar (false).
+ * @param {string} userPhoto - URL foto profil user.
+ */
+export const playerJoinRoom = async (
+  topicID,
+  gameID,
+  roomID,
+  user,
+  isJoining,
+  userPhoto
+) => {
+  if (roomID === "room5") return; // Room AI tidak memerlukan join/leave manual
 
   if (!topicID || !gameID || !roomID || !user?.uid) {
     console.error("Missing required parameters in roomsParticipation");
@@ -42,21 +48,23 @@ export const roomsParticipation = async (topicID, gameID, roomID, user, isJoinin
     if (roomSnapshot.val().isSinglePlayer) return;
 
     if (!isJoining) {
-      await remove(ref(database, playerPath));  // Hapus pemain dari room
+      // Menghapus pemain jika keluar
+      await remove(ref(database, playerPath));
     } else {
+      // Gunakan getRoomPlayerCount  untuk memvalidasi kapasitas room sebelum menambahkan pemain
       const capacity = roomSnapshot.val().capacity || 4;
-      const playersSnapshot = await get(ref(database, `${roomPath}/players`));
-      const playerCount = playersSnapshot.exists() ? Object.keys(playersSnapshot.val()).length : 0;
+      const playerCount = await getRoomPlayerCount(topicID, gameID, roomID);
 
       if (playerCount >= capacity) {
         throw new Error("Room is full");
       }
 
+      // Menambahkan data pemain ke room
       await set(ref(database, playerPath), {
         uid: user.uid,
         displayName: user.displayName,
         photoURL: userPhoto || "",
-        joinedAt: Date.now()
+        joinedAt: Date.now(),
       });
 
       await set(ref(database, `${roomPath}/currentPlayers`), playerCount + 1);
@@ -67,68 +75,109 @@ export const roomsParticipation = async (topicID, gameID, roomID, user, isJoinin
   }
 };
 
-export const subscribeToPlayers = (topicID, gameID, roomID, onPlayersUpdate) => {
+/**
+ * Mengawasi perubahan data pemain dalam room secara real-time.
+ * @param {string} topicID - ID topik.
+ * @param {string} gameID - ID game.
+ * @param {string} roomID - ID room.
+ * @param {function} onPlayersUpdate - Callback untuk mengirim data pemain.
+ */
+export const fetchPlayer = (topicID, gameID, roomID, onPlayersUpdate) => {
   if (roomID === "room5") return () => {};
 
-  if (!topicID || !gameID || !roomID || typeof onPlayersUpdate !== 'function') {
-    console.error("Invalid parameters for subscribeToPlayers");
+  if (!topicID || !gameID || !roomID || typeof onPlayersUpdate !== "function") {
+    console.error("Invalid parameters for fetch data player");
     return () => {};
   }
 
-  const playersRef = ref(database, `rooms/${topicID}/${gameID}/${roomID}/players`);
+  const playersRef = ref(
+    database,
+    `rooms/${topicID}/${gameID}/${roomID}/players`
+  );
   return onValue(playersRef, (snapshot) => {
     const playersData = snapshot.val() || {};
     onPlayersUpdate(Object.values(playersData));
   });
 };
 
+/**
+ * Mengambil data semua pemain yang ada dalam room.
+ * @param {string} topicID - ID topik.
+ * @param {string} gameID - ID game.
+ * @param {string} roomID - ID room.
+ * @returns {Array} - Daftar pemain dalam room.
+ */
+export const getCurrentPlayers = async (topicID, gameID, roomID) => {
+  if (roomID === "room5") return [];
+
+  const playersRef = ref(
+    database,
+    `rooms/${topicID}/${gameID}/${roomID}/players`
+  );
+  const snapshot = await get(playersRef);
+  return snapshot.val() ? Object.values(snapshot.val()) : [];
+};
+
+/**
+ * Sinkronisasi jumlah pemain dalam room ke database.
+ * @param {string} topicID - ID topik.
+ * @param {string} gameID - ID game.
+ * @param {string} roomID - ID room.
+ * @returns {number} - Jumlah pemain dalam room.
+ */
 export const syncCurrentPlayers = async (topicID, gameID, roomID) => {
   if (!topicID || !gameID || !roomID || roomID === "room5") return;
 
   try {
-    const playersRef = ref(database, `rooms/${topicID}/${gameID}/${roomID}/players`);
-    const snapshot = await get(playersRef);
-    const playerCount = snapshot.exists() ? Object.keys(snapshot.val()).length : 0;
-    
-    await set(ref(database, `rooms/${topicID}/${gameID}/${roomID}/currentPlayers`), playerCount);
+    // Menggunakan getRoomPlayerCount  untuk mendapatkan jumlah pemain
+    const playerCount = await getRoomPlayerCount(topicID, gameID, roomID);
+
+    await set(
+      ref(database, `rooms/${topicID}/${gameID}/${roomID}/currentPlayers`),
+      playerCount
+    );
     return playerCount;
   } catch (error) {
     console.error("Error syncing players:", error);
   }
 };
 
-export const resetRoom = async (topicID, gameID, roomID) => {
-  if (!topicID || !gameID || !roomID || roomID === "room5") return;
-
-  try {
-    const playersRef = ref(database, `rooms/${topicID}/${gameID}/${roomID}/players`);
-    const snapshot = await get(playersRef);
-    const playerCount = snapshot.exists() ? Object.keys(snapshot.val()).length : 0;
-
-    if (playerCount === 0) {
-      const updates = {
-        [`rooms/${topicID}/${gameID}/${roomID}/gameStatus`]: null,
-        [`rooms/${topicID}/${gameID}/${roomID}/currentPlayers`]: 0,
-        [`rooms/${topicID}/${gameID}/${roomID}/chatMessages`]: null  // Tambahkan ini
-      };
-      await update(ref(database), updates);
-      console.log("Room reset: game status, chat messages, and current players cleared");
-    }
-  } catch (error) {
-    console.error("Error resetting game status:", error);
-  }
-};
-
-
+/**
+ * Mengelola keluarnya pemain dari room dan memperbarui currentPlayers dengan transaksi.
+ * @param {string} topicID - ID topik.
+ * @param {string} gameID - ID game.
+ * @param {string} roomID - ID room.
+ * @param {object} user - Objek user yang berisi UID dan detail lainnya.
+ */
 export const playerLeaveRoom = async (topicID, gameID, roomID, user) => {
-  if (!topicID || !gameID || !roomID || !user?.uid || roomID === "room5") return;
+  if (!topicID || !gameID || !roomID || !user?.uid || roomID === "room5")
+    return;
+
+  const playerPath = `rooms/${topicID}/${gameID}/${roomID}/players/${user.uid}`;
+  const currentPlayersRef = ref(
+    database,
+    `rooms/${topicID}/${gameID}/${roomID}/currentPlayers`
+  );
 
   try {
-    // Hapus pemain dari room
-    await roomsParticipation(topicID, gameID, roomID, user, false);
-    
-    // Periksa apakah perlu reset room setelah pemain terakhir keluar
-    await resetRoom(topicID, gameID, roomID);
+    // Hapus pemain dari daftar pemain
+    await remove(ref(database, playerPath));
+
+    // Perbarui currentPlayers menggunakan transaksi
+    await runTransaction(currentPlayersRef, (currentCount) => {
+      if (currentCount) {
+        return currentCount - 1;
+      } else {
+        return 0;
+      }
+    });
+
+    // Dapatkan jumlah pemain yang tersisa
+    const playerCount = await getRoomPlayerCount(topicID, gameID, roomID);
+    if (playerCount === 0) {
+      // Reset room jika pemain terakhir keluar
+      await resetRoom(topicID, gameID, roomID);
+    }
   } catch (error) {
     console.error("Error in playerLeaveRoom:", error);
   }
